cuts calculated

+++ Models/LockSpecialist.cs
@@ -5,7 +5,7 @@ namespace HeistPart2 {
   public class LockSpecialist : IRobber {
     public string Name { get; set; }
     public int SkillLevel { get; set; }
-    public int PercentageCut { get; set; }
+    public double PercentageCut { get; set; }
 
     public void PerformSkill (Bank bank) {
       bank.VaultScore -= SkillLevel;
